Bem.Cfg.Cfg

Description

general configuration

The settings are described regarding an entity as shown, that is, as a string.

Documentation

data Cfg

the available settings (along with their default values)

Constructors

Cfg

Fields

_elemSep :: Sep

a separator between a block and its element inside a class ("__")

_modSep :: Sep

a separator between a non-modififer entity, a modifier, and an optional value of the latter inside a class ("_")

_partSep :: Sep

a separator between entity parts inside a class ("-")

_partsAreCptled :: Bool

whether to capitalise entity parts inside a class (False)

data FixedCfg

the safety wrapper around configuration whose value can only be created using the fix function

type Sep = String

an entity part separator inside a class

defCfg :: Cfg

the default settings

fix :: Cfg -> FixedCfg

If a particular combination of setting values is occuried, then some settings are defaulted back to avoid malformed custom configurations.

unfix :: FixedCfg -> Cfg

Obtain an actual configuration after fixing it.