Bem.Cls.Gen.Gen

Description

Generate a class of a combination of some entities from a scheme with the default decorations.

Documentation

genBlk :: FromFullBlk b (FromFullElem b Class)

Generate a class of a block and element with their modifiers.

genElem :: FromFullElem b Class

Generate a class of a element with its modifiers.