Bem.Cls.Gen.Intr

Description

low-level API for class generation with custom decorations

Documentation

data Ent (isMod :: Bool) where

the GADT of wrappers allowing entities to be of the same type

Constructors

Blk :: Show (b e m) => b (e :: Type -> Type) m -> Ent False

BoolMod :: Show m => Ent False -> m -> Ent True

Elem :: (Show (b e bm), Show (e em), Show em) => b e bm -> e em -> Ent False

Mod :: Show m => Ent False -> m -> Ent True

decor :: Ent isMod -> Reader FixedCfg StredEnt

Decorate an entity, wrapped into a corresponding Ent data constructor, using custom decorations

str :: Ent isMod -> Reader FixedCfg FullStredEnt

the topmost helper function gathering all the other functions to stringify a given wrapped entity

type FullStredEnt = String

a final decorated entity with everyting needed attached

type StredEnt = String

a decorated entity

genBlk :: FromBlkElem (Reader FixedCfg Class)

Generate a class of a block and element with their modifiers, using custom decorations.

genElem :: FromFullElem b (Reader FixedCfg Class)

Generate a class of a element with its modifiers, using custom decorations.