Bem.Utl.Utl

Description

high-level extra utilities

Documentation

data Gens

the configurable extra class generators

Constructors

Gens

Fields

_genBlk :: FromBlkElem Class

_genElem :: forall b. FromFullElem b Class

_genNoBlkModsBlk :: forall b. FromBlk b (FromFullElem b Class)

_genNoElemModsBlk :: FromBlkNoModsElem Class

_genNoModsBlk :: FromNoModsBlkNoModsElem Class

_genNoModsElem :: forall b. FromElem b Class

data NoElem m

Denote absent elements.

data NoMod

Denote absent modifiers.

genNoBlkModsBlk :: FromNoModsBlkElem Class

Generate a class of a block without its modifiers and full element.

genNoElemModsBlk :: FromBlkNoModsElem Class

Generate a class of a block and element that is without its modifiers.

genNoModsBlk :: FromNoModsBlkNoModsElem Class

Generate a class of a block and element without any modifiers.

genNoModsElem :: FromElem b Class

Generate a class of an element that is without its modifiers.

decorSingleton :: Show (b e m) => b (e :: Type -> Type) m -> Class

Decorate a single block.

init :: Cfg -> Gens

Initialise the utility configurable class generators with a configuration.