Bem.Cls.Gen.Gen

Description

Generate a class of some entities combination from a scheme using the default decorations.

Documentation

genBlkElem :: FromBlkElem Class

Generate a class of a block and element along with their modifiers.

genElem :: FromFullElem b Class

Generate a class of a element along with its modifiers.